Pt.1, adding product listing support to transaction classes
As a way of grabbing information from Transaction classes and giving
them to Shopping cart, a ProductListing object has been added to
Transaction objects. Future additions will include a listing of
ProductEntity objects (if needed) and the ability to populate the lists
from the SQL worksheet using transaction2 and columns with the same
TransID. (Time spent: Appx. 2.5 hours, would have spent longer if not
for class.)